Added support for Windows build
* Removed source of non-determinism which caused different indices
assigned to states across Windows and Linux (replacement of
std::unordered_map with std::map)
* Type casts because of signed/unsigned or 8-byte to 4-byte int
conversion
* Build system for downloading re2 and fmt had to be modified to support
MSVC build
* hash_combine() tests were reduced down to comparison of
equality/inequality with other hashes
* New structure SymbolLess<T> for comparing two Symbol<T> pointers in
std::less<T> way
* Added missing std::optional wherever possible

+++ include/pog/grammar.h
@@ -97,7 +97,7 @@ public:
 		if (auto itr = _name_to_symbol.find(name); itr != _name_to_symbol.end())
 			return itr->second;
 
-		_symbols.push_back(std::make_unique<SymbolType>(_symbols.size(), kind, name));
+		_symbols.push_back(std::make_unique<SymbolType>(static_cast<std::uint32_t>(_symbols.size()), kind, name));
 		_name_to_symbol.emplace(_symbols.back()->get_name(), _symbols.back().get());
 		return _symbols.back().get();
 	}
@@ -105,7 +105,7 @@ public:
 	template <typename CallbackT>
 	RuleType* add_rule(const SymbolType* lhs, const std::vector<const SymbolType*>& rhs, CallbackT&& action)
 	{
-		_rules.push_back(std::make_unique<RuleType>(_rules.size(), lhs, rhs, std::forward<CallbackT>(action)));
+		_rules.push_back(std::make_unique<RuleType>(static_cast<std::uint32_t>(_rules.size()), lhs, rhs, std::forward<CallbackT>(action)));
 		return _rules.back().get();
 	}
